Some Trick jobs are labeled "simobject.classPtr->Foo()" trick_dp barfs on the "->" string
The "->" string interferes with some code in quick plot trying to split array ranges into
individual variables.  The old code only tested for the '-' character.  Changed these
checks to regular expressions that only match '-' if it is not followed by '>'.  Also
converted tabs to spaces throughout the file, hence the big diff.

refs #51