Update react-based Trick Dashboard (#900)
Fast-forwarding my app to the latest version to make it available and minimize integration problems later on. The app still requires several features to be implemented such as a Search feature for TV.

Currently, the process for using the app is as follows:

Install Node.js. https://nodejs.org/en/ or alternatively find a repository such as NodeSource
```cd trick/trick_source/web/mockup```
```npm i```
```npm run build```
Copy the contents of the build directory to your www directory for your sim